improve: aql style
